+++ CriticalMass/SettingsProjectLinkTableViewCell.swift
@@ -66,19 +66,22 @@ extension SettingsProjectLinkTableViewCell {
         let title: String
         let detail: String
         let actionTitle: String
+        let userInputLabels: [String]
         let image: UIImage
 
         static let github = CellConfiguration(
             title: L10n.settingsOpenSourceTitle,
             detail: L10n.settingsOpenSourceDetail,
             actionTitle: L10n.settingsOpenSourceAction.uppercased(),
+            userInputLabels: [L10n.settingsOpenSourceAction],
             image: UIImage(named: "GithubBanner")!
         )
 
         static let criticalMassDotIn = CellConfiguration(
             title: L10n.settingsCriticalMassDotInTitle,
             detail: L10n.settingsCriticalMassDotInDetail,
-            actionTitle: L10n.settingsOpenSourceAction.uppercased(),
+            actionTitle: L10n.settingsCriticalMassDotInAction.uppercased(),
+            userInputLabels: [L10n.settingsCriticalMassDotInAction],
             image: UIImage(named: "banner cm")!
         )
     }