+++ CHANGELOG.md
@@ -17,6 +17,7 @@ Changelog for Critical Maps iOS
 - Fix Wrong displayed timestamps for tweets if the App was active for a while
 - Fix Dark Mode Color Behaviour in Action Indicator (in Rules, Settings) and Separator (in Settings)
 - Fix missing information for VoiceOver user for the navigationbar buttons
+- Fix VoiceControl labels on each view
 
 ### Changed